Add files via upload
This is the clean version of the fifa21 raw data v2.csv. It took me 3 days to clean this dataset. Best regards to Pastyasamoah his work was of great help.
Big thanks to Chinoso and her group for bring up this #datacleaningchallenge it added a lot to my knowledge base.
This work maybe far from perfect but we learn everyday.